refactor(ai): unify ZIP parsing in resource parsers
Move bundled Skill and AgentSpec seed archive parsing into their respective ZipParser classes while preserving bootstrap package metadata and limits.
