+++ ai/src/test/java/com/alibaba/nacos/ai/utils/SkillZipParserSeedArchiveTest.java
@@ -39,7 +39,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class SkillSeedArchiveReaderTest {
+class SkillZipParserSeedArchiveTest {
     
     @Test
     void shouldBuildStandaloneSkillZipWithTopLevelResources() throws Exception {
@@ -48,8 +48,8 @@ class SkillSeedArchiveReaderTest {
             new ArchiveEntry("vendor/demo-skill/LICENSE.txt", "license"),
             new ArchiveEntry("vendor/demo-skill/references/guide.md", "guide"));
         
-        List<SkillSeedArchiveReader.SkillPackage> actual =
-            SkillSeedArchiveReader.read(new ByteArrayInputStream(archive));
+        List<SkillZipParser.SkillPackage> actual =
+            SkillZipParser.parseSkillPackagesFromZip(new ByteArrayInputStream(archive));
         
         assertEquals(1, actual.size());
         assertEquals("demo-skill", actual.get(0).getSkillName());
@@ -67,8 +67,8 @@ class SkillSeedArchiveReaderTest {
             new ArchiveEntry("source-a/demo/SKILL.md", buildSkillMarkdown("same-skill")),
             new ArchiveEntry("source-b/demo/SKILL.md", buildSkillMarkdown("same-skill")));
         
-        List<SkillSeedArchiveReader.SkillPackage> actual =
-            SkillSeedArchiveReader.read(new ByteArrayInputStream(archive));
+        List<SkillZipParser.SkillPackage> actual =
+            SkillZipParser.parseSkillPackagesFromZip(new ByteArrayInputStream(archive));
         
         assertEquals(1, actual.size());
         assertEquals("same-skill", actual.get(0).getSkillName());
@@ -78,8 +78,8 @@ class SkillSeedArchiveReaderTest {
     void shouldParseNestedFromPath() throws Exception {
         byte[] archive = buildArchive(new ArchiveEntry("github.com/nacos/find-skills/SKILL.md",
             buildSkillMarkdown("find-skills")));
-        List<SkillSeedArchiveReader.SkillPackage> actual =
-            SkillSeedArchiveReader.read(new ByteArrayInputStream(archive));
+        List<SkillZipParser.SkillPackage> actual =
+            SkillZipParser.parseSkillPackagesFromZip(new ByteArrayInputStream(archive));
         assertEquals(1, actual.size());
         assertEquals("github.com/nacos", actual.get(0).getFrom());
     }
@@ -92,7 +92,7 @@ class SkillSeedArchiveReaderTest {
             new ArchiveEntry("vendor/demo-skill/assets/demo.txt", "demo"));
         
         IOException exception = assertThrows(IOException.class,
-            () -> SkillSeedArchiveReader.read(new ByteArrayInputStream(archive), 2,
+            () -> SkillZipParser.parseSkillPackagesFromZip(new ByteArrayInputStream(archive), 2,
                 1024 * 1024));
         
         assertTrue(exception.getMessage().contains("too many entries"));
@@ -107,7 +107,8 @@ class SkillSeedArchiveReaderTest {
         long maxBytes = skillMarkdown.getBytes(StandardCharsets.UTF_8).length + 5L;
         
         IOException exception = assertThrows(IOException.class,
-            () -> SkillSeedArchiveReader.read(new ByteArrayInputStream(archive), 10, maxBytes));
+            () -> SkillZipParser.parseSkillPackagesFromZip(new ByteArrayInputStream(archive), 10,
+                maxBytes));
         
         assertTrue(exception.getMessage().contains("decompressed size exceeds limit"));
     }
@@ -120,7 +121,7 @@ class SkillSeedArchiveReaderTest {
             new ArchiveEntry("./vendor/demo-skill/SKILL.md", skillMarkdown));
         
         IOException exception = assertThrows(IOException.class,
-            () -> SkillSeedArchiveReader.read(new ByteArrayInputStream(archive), 10,
+            () -> SkillZipParser.parseSkillPackagesFromZip(new ByteArrayInputStream(archive), 10,
                 1024 * 1024));
         
         assertTrue(exception.getMessage().contains("duplicate entry path"));
@@ -132,12 +133,12 @@ class SkillSeedArchiveReaderTest {
         Assumptions.assumeTrue(resource.exists(),
             "bootstrap/skills-data.zip is not bundled in this test runtime");
         try (InputStream inputStream = resource.getInputStream()) {
-            List<SkillSeedArchiveReader.SkillPackage> actual =
-                SkillSeedArchiveReader.read(inputStream);
+            List<SkillZipParser.SkillPackage> actual =
+                SkillZipParser.parseSkillPackagesFromZip(inputStream);
             
             assertEquals(139, actual.size());
             Set<String> skillNames = new HashSet<>();
-            for (SkillSeedArchiveReader.SkillPackage each : actual) {
+            for (SkillZipParser.SkillPackage each : actual) {
                 skillNames.add(each.getSkillName());
                 Skill skill = SkillZipParser.parseSkillFromZip(each.getZipBytes(), "public");
                 assertNotNull(skill);
